checkpoint: remove limitation that avoid m-mode
Because the multi-core synchronization constraint will confilit
with RISC-V IPI implementation: Core A reach simulation point and
waits for core B to 1) reach the same point and 2) leave m-mode.
However, if a RISC-V IPI raises from core B, core B will stay in
m-mode to WAIT core A to response! Then A and B deadlock.